// src/app/gameItem.interface.ts
export interface GameItemInterface {
  id: number;
  name: string;
  year: number;
  developer: string;
  description: string;
  website: string;
  game_engines: string[];
  player_modes: string[];
  platforms: string[];
  pov: string[];
  genres: string[];
  themes: string[];
  keywords: string[];
  img: string;
  playtime: number,
  rating: number;
}